# internal/server/openapi/openapi-admin.yaml
openapi: 3.1.0
info:
  title: Firefly Admin API
  # Placeholder only. The running binary replaces this with its build version
  # at startup (server.SetBuildVersion); do not hand-edit it to track a release.
  version: "0.0.0-dev"
  description: >
    Administrative surface for the Firefly LLM gateway dashboard. Every
    endpoint requires an authenticated operator: either an active dashboard
    session token (obtained via `POST /api/auth/login`) or the configured
    admin bearer token, presented as `Authorization: Bearer <token>`.

    This document covers tenant lifecycle management under `/api/tenants`
    and provider/credential-pool administration under `/api/providers` and
    `/api/keys`. Tenant mutations persist through the same validate-then-store
    pipeline as the bulk settings endpoint (Turso when configured, otherwise
    the config files) and hot-swap the routing snapshot, so changes take effect
    immediately without a restart. Provider and key records live only in the
    Turso database and are refused with 503 when no store is configured. The
    public `/v1/*` surface is documented separately at `/api/openapi.yaml`.

    Credential discipline: stored keys are never returned. Key reads carry a
    masked `api_key_hint` only, and because that hint is the sole credential
    shape this surface ever shows, a secret in mask shape is refused on writes
    instead of being stored as a routable credential. `account_metadata` — the
...
  contact:
    name: Firefly
...
servers:
...
security:
  - AdminAuth: []
paths:
  /api/tenants:
    get:
      tags: [Tenants]
      summary: List tenants
      description: >
        Returns the full tenant catalog from the authoritative configuration
        store (Turso when configured, otherwise the config files). Tenant API
        keys are returned in plaintext, like `GET /api/settings`, because they
        are working credentials the operator must be able to copy.
      operationId: listTenants
      responses:
        "200":
          description: The tenant catalog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      tags: [Tenants]
      summary: Create tenant
      description: >
        Appends a tenant to the catalog. The name must be unique; duplicate
        names return 409. When `api_key` is omitted a fresh gateway key
        (`sk-gw-` plus random hex) is generated automatically, and `key_hash`
        is derived from the plaintext key when absent. The full catalog is
        re-validated before persisting: an unknown `allowed_models` entry or a
        missing `credential_ref` environment variable fails the request with
        400 and changes nothing.
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantCreate"
      responses:
        "201":
          description: Tenant created and live.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/tenants/{name}:
    get:
      tags: [Tenants]
      summary: Retrieve a tenant
      description: Returns one tenant by display name.
      operationId: getTenant
      parameters:
        - $ref: "#/components/parameters/TenantName"
      responses:
        "200":
          description: The tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    put:
      tags: [Tenants]
      summary: Update tenant
      description: >
        Overlays the operator-managed fields of an existing tenant; absent or
        zero-valued fields keep their stored values, so partial edits never
        silently clear quota or access. Credential rotation happens only when
        the body carries an explicit, unmasked `api_key` that differs from the
        stored one — omitting `api_key` or echoing the masked placeholder from
        a previous GET preserves the current credential. `used_tokens` is
        metered state owned by the traffic pipeline and the topup endpoint; it
        is never taken from the request.
      operationId: updateTenant
      parameters:
        - $ref: "#/components/parameters/TenantName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Tenant updated and live.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Tenants]
      summary: Delete tenant
      description: >
        Removes the tenant from the catalog, including deleting the last
        remaining tenant. Subsequent requests authenticated with the tenant's
        key fail with 401.
      operationId: deleteTenant
      parameters:
        - $ref: "#/components/parameters/TenantName"
      responses:
        "200":
          description: Tenant deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantDeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/tenants/topup:
    post:
      tags: [Tenants]
      summary: Top up tenant balance
      description: >
        Adds tokens, extends the expiry window, and/or resets consumed usage
        for a tenant identified by `api_key` or `tenant_name`. An
        `exhausted` or `expired` tenant returns to `active`. Persisted to the
        database when Turso is configured.
      operationId: topupTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantTopupRequest"
      responses:
        "200":
          description: Top-up applied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantTopupResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/providers:
    get:
      tags: [Providers]
      summary: List providers
      description: >
        Returns every provider record, active or inactive, ordered by name,
        each with the count of keys currently routable through the same
        predicate the data plane uses. Unlike the legacy
        `GET /api/turso/providers`, a storage failure is surfaced as an error
        instead of degrading to an empty list.
      operationId: listProviders
      responses:
        "200":
          description: The provider catalog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      tags: [Providers]
      summary: Create provider
      description: >
        Inserts one provider. `name` and `base_url` are required, and the name
        must be unique because it is the natural key the harvester syncs
        against — a duplicate returns 409 rather than silently capturing the
        existing pool. The catalog revision is bumped and an immediate reload
        is attempted, so the record is visible to routing before the response
        returns.
      operationId: createProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderCreate"
      responses:
        "201":
          description: Provider created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /api/providers/{id}:
    get:
      tags: [Providers]
      summary: Retrieve a provider
      description: Returns one provider by numeric id, active or not.
      operationId: getProvider
      parameters:
        - $ref: "#/components/parameters/ProviderID"
      responses:
        "200":
          description: The provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    put:
      tags: [Providers]
      summary: Update provider
      description: >
        Partial update: `base_url`, `description`, and `is_active` are applied
        when present, and an absent field keeps its stored value. The name is
        not editable (providers are matched by name, so a rename would orphan
        the credential pool); sending a different name is a 400. An empty body
        is rejected as a client bug rather than treated as a no-op.
      operationId: updateProvider
      parameters:
        - $ref: "#/components/parameters/ProviderID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderUpdate"
      responses:
        "200":
          description: Provider updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderMutationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [Providers]
      summary: Delete provider
      description: >
        Hard-deletes the provider together with its keys and their bound
        credentials (the operator-only right; the harvester can only upsert and
        deactivate). Refused with 409 while any upstream still references the
        provider, because the delete would silently detach that upstream's
        credential pool — detach or delete the upstreams first.
      operationId: deleteProvider
      parameters:
        - $ref: "#/components/parameters/ProviderID"
      responses:
        "200":
          description: Provider and its keys deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderDeleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/pricing:
    get:
      tags: [Pricing]
      summary: List the local price sheet
      description: >
        Returns every entry in the local price sheet, each annotated with a
        `used` flag that says whether the gateway's own catalog actually routes
        a model matching that key. Prices are integer micro-USD per 1M tokens.
      operationId: listPricing
      responses:
        "200":
          description: The price sheet.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/PricingEntryView"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/pricing/catalog/refresh:
    post:
      tags: [Pricing]
      summary: Refresh the models.dev catalog cache
      description: >
        Force-fetches the catalog and replaces the in-memory cache, so a
        stale picker can be renewed without waiting for a restart. Like the
        browse endpoint, this never mutates the persisted price sheet.
      operationId: refreshPricingCatalog
      responses:
        "200":
          description: The freshly fetched catalog, unfiltered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingCatalogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          description: The models.dev catalog could not be fetched.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingCatalogResponse"
components:
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      description: >
        Dashboard session token (from `POST /api/auth/login`) or the
        configured admin bearer token. A missing or invalid token returns 401.
...
  schemas:
...
    ProviderResponse:
      type: object
      required: [provider]
      properties:
        provider:
          $ref: "#/components/schemas/Provider"
...